[gitlab] don't clone industrial_ci into target folder
Gitlab CI runs in the target source folder by default.
If industrial_ci gets cloned there, it might be part of the build.
In case of testing industrial_ci with industrial_ci this leads to duplicated package names.
In addition Gitlab's docker-in-docker service syncs only files in /builds/$CI_PROJECT_NAMESPACE.
That's why it is has to be placed in ..