[gitlab] fix prerelease tests
TMPDIR is used as the parent directory in mktemp.
Because of limitation of Gitlab's docker-in-docker service, the temporary files for the pre-release tests have to be placed underneath /builds/$CI_PROJECT_NAMESPACE
Furthermore, the directory needs to exist to make mktemp work. Luckily Gitlab CI creates ${CI_PROJECT_DIR}.tmp